# [Refactor:System] General performance improvements