--- a/site/app/models/AbstractModel.php
+++ b/site/app/models/AbstractModel.php
@@ -15,7 +15,13 @@ use app\libraries\Utils;
  */
 abstract class AbstractModel {
 
+    const CALL_SET = 1;
+    const CALL_GET = 2;
+    const CALL_IS = 3;
+
     protected static $properties = [];
+
+    protected static $cache = [];
 
     /** @var Core */
     protected $core;
@@ -125,8 +131,29 @@ abstract class AbstractModel {
      */
     public function __call($name, $arguments) {
         $class_name = get_class($this);
-        if (Utils::startsWith($name, "set")) {
-            $property_name = $this->convertName($name);
+        if (!isset(static::$cache[$class_name][$name])) {
+            if (Utils::startsWith($name, "set")) {
+                static::$cache[$class_name][$name] = [
+                    "function_type" => AbstractModel::CALL_SET,
+                    "property_name" => $this->convertName($name)
+                ];
+            }
+            elseif (Utils::startsWith($name, "get")) {
+                static::$cache[$class_name][$name] = [
+                    "function_type" => AbstractModel::CALL_GET,
+                    "property_name" => $this->convertName($name)
+                ];
+            }
+            elseif (Utils::startsWith($name, "is")) {
+                static::$cache[$class_name][$name] = [
+                    "function_type" => AbstractModel::CALL_IS,
+                    "property_name" => $this->convertName($name, 2)
+                ];
+            }
+        }
+        $function_type = static::$cache[$class_name][$name]["function_type"] ?? 0;
+        $property_name = static::$cache[$class_name][$name]["property_name"] ?? "";
+        if ($function_type === AbstractModel::CALL_SET) {
             $value = $arguments[0];
             if (isset(static::$properties[$class_name][$property_name])) {
                 if (isset(static::$properties[$class_name][$property_name]['read_only'])) {
@@ -156,8 +183,7 @@ abstract class AbstractModel {
                 return null;
             }
         }
-        elseif (Utils::startsWith($name, "get")) {
-            $property_name = $this->convertName($name);
+        elseif ($function_type === AbstractModel::CALL_GET) {
             if (
                 isset(static::$properties[$class_name][$property_name])
                 && isset(static::$properties[$class_name][$property_name]['write_only'])
@@ -167,8 +193,7 @@ abstract class AbstractModel {
 
             return $this->$property_name;
         }
-        elseif (Utils::startsWith($name, "is")) {
-            $property_name = $this->convertName($name, 2);
+        elseif ($function_type === AbstractModel::CALL_IS) {
             if (
                 isset(static::$properties[$class_name][$property_name])
                 && isset(static::$properties[$class_name][$property_name]['write_only'])