### What is the new behavior?
Performance improvements on AbstractModel, GradingOrder, and Component.
While this is aimed at improving performance on TA grading pages, it also will likely improve performance on other pages as well (the improvements to AbstractModel should improve performance on many pages).

On AbstractModel, a cache (not persistent) is added in to cache function types/name conversions, specifically to avoid redundant calls to Utils::startsWith and AbstractModel's convertName method.

On GradingOrder, getSortedGradedGradeables search is improved to avoid unnecessarily using getSubmitterIndex each time.

On Component, the call to getPrecision for a gradeable is moved outside of the for loop to avoid redundant calls.

Test results for improvements:

Grading index for gradeable "grading_homework" inside of the "sample" course as user "submitty-admin" on view all sections:
Cumulative load time before improvements: 2900 ms
Cumulative load time after improvements: 1168 ms

Grading page for user "aphacker" on gradeable "grading_homework" in the "sample" course as user "submitty-admin":
Cumulative load time before improvements: 1440 ms
Cumulative load time after improvements: 784 ms

Tested using Xdebug profiler, with debug mode disabled for Submitty and caches loaded for pages (each page loaded once after reload site before profiler was run); read profiler info with WinCacheGrind.

### Other information?
On the grading page for TA grading, more improvements can be made with improving late day calculations.